feat(grimmory): replace booklore with grimmory
Booklore has been replaced by Grimmory as the book library service.
Updates all service config, Ansible vars, Terraform/ACL references,
tests, and documentation to use grimmory throughout.
