+++ ansible/roles/nexus/templates/access-rules.yml.j2
@@ -76,7 +76,7 @@ services:
     groups: [admins]
     description: "Document management"
 
-  booklore:
+  grimmory:
     groups: [admins]
     description: "Book library"
