+++ config/presets.yml
@@ -16,7 +16,7 @@ home:
     - backups
     - sure
     - paperless
-    - booklore
+    - grimmory
     - foundryvtt
     - jellyfin
     - transmission
