--- services/booklore/docker-compose.yml
@@ -1,68 +1,0 @@
-services:
-  booklore:
-    image: booklore/booklore:latest
-    container_name: booklore
-    restart: unless-stopped
-    deploy:
-      resources:
-        limits:
-          memory: 512M
-        reservations:
-          memory: 128M
-    networks:
-      - nexus
-      - booklore
-    volumes:
-      - ${NEXUS_DATA_DIRECTORY}/Config/booklore/data:/app/data
-      - ${BOOKLORE_BOOKS_DIRECTORY}:/books
-      - ${NEXUS_DATA_DIRECTORY}/Config/booklore/bookdrop:/bookdrop
-    environment:
-      DATABASE_URL: jdbc:mariadb://booklore-db:3306/booklore
-      DATABASE_USERNAME: ${BOOKLORE_MYSQL_USER}
-      DATABASE_PASSWORD: ${BOOKLORE_MYSQL_PASSWORD}
-      DISK_TYPE: LOCAL
-      TZ: ${TZ}
-    depends_on:
-      booklore-db:
-        condition: service_healthy
-    labels:
-      - "traefik.enable=true"
-      - "traefik.docker.network=nexus"
-      - "traefik.http.services.booklore.loadbalancer.server.port=6060"
-      - "traefik.http.routers.booklore.rule=Host(`booklore.${NEXUS_DOMAIN}`)"
-      - "traefik.http.routers.booklore.entrypoints=https"
-      - "traefik.http.routers.booklore.middlewares=tailscale-chain@file"
-      - "traefik.http.routers.booklore.tls=true"
-      - "traefik.http.routers.booklore.tls.certresolver=certchallenge"
-
-  booklore-db:
-    image: lscr.io/linuxserver/mariadb:11.4.9
-    container_name: booklore-db
-    restart: unless-stopped
-    deploy:
-      resources:
-        limits:
-          memory: 256M
-        reservations:
-          memory: 64M
-    networks:
-      - booklore
-    volumes:
-      - ${NEXUS_DATA_DIRECTORY}/Config/booklore/mariadb:/config
-    environment:
-      MYSQL_ROOT_PASSWORD: ${BOOKLORE_MYSQL_ROOT_PASSWORD}
-      MYSQL_DATABASE: booklore
-      MYSQL_USER: ${BOOKLORE_MYSQL_USER}
-      MYSQL_PASSWORD: ${BOOKLORE_MYSQL_PASSWORD}
-    healthcheck:
-      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-p$$MYSQL_ROOT_PASSWORD"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-networks:
-  nexus:
-    external: true
-  booklore:
-    name: nexus_booklore
-    driver: bridge
